fixed warnings on Mac; added r: width param to display strings; allowed -lo(g)s short Name on command line params

+++ src/libs/etherlib/accountwatch.cpp
@@ -320,7 +320,7 @@ bool CAccountWatch::isTransactionOfInterest(CTransaction *trans, uint32_t nSigs,
     //
     // TODO(tjayrush): We should do a 'deep trace' here (or when the block is first read)
     // to see if there was a 'call,' to our address.
-    for (uint64_t i = 0 ; i < trans->receipt.logs.getCount() ; i++) {
+    for (uint32_t i = 0 ; i < trans->receipt.logs.getCount() ; i++) {
         SFString acc = address;
         CLogEntry *l = reinterpret_cast<CLogEntry *>(&trans->receipt.logs[i]);
         if (l->address.Contains(acc)) {
@@ -352,7 +352,7 @@ bool CAccountWatch::isTransactionOfInterest(CTransaction *trans, uint32_t nSigs,
                 }
 
                 // ...or the address is in the indexed topics or data
-                for (uint64_t j = 1 ; j < l->topics.getCount() ; j++) {
+                for (uint32_t j = 1 ; j < l->topics.getCount() ; j++) {
                     SFHash tHash = fromTopic(l->topics[j]);
                     if (tHash % acc) {
                         trans->isInternalTx = true;
@@ -366,7 +366,7 @@ bool CAccountWatch::isTransactionOfInterest(CTransaction *trans, uint32_t nSigs,
     if (deepScan) {
         CTraceArray traces;
         getTraces(traces, trans->hash);
-        for (uint64_t i = 0 ; i < traces.getCount() ; i++) {
+        for (uint32_t i = 0 ; i < traces.getCount() ; i++) {
             CTraceAction *action = (CTraceAction*)&(traces[i].action);
             if (action->to % address || action->from % address || action->address % address || action->refundAddress % address) {
                 trans->isInternalTx = true;
